[react-devtools-shared] extract getDispatcherRef into a shared module
Move getDispatcherRef out of backend/fiber/renderer.js into
backend/shared/DevToolsReactDispatcher.js and import it back. This lets other
consumers (e.g. react-devtools-facade) reuse the renderer-dispatcher adapter
without importing the whole renderer module.
